+++ src/Accounts/Authentication/Authentication/TokenCache/PowerShellTokenCacheProvider.cs
@@ -12,29 +12,29 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
 using Azure.Identity;
 
 using Hyak.Common;
 
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.Commands.Common.Authentication.Utilities;
-using Microsoft.Azure.Commands.Shared.Config;
 using Microsoft.Azure.Internal.Subscriptions;
 using Microsoft.Azure.Internal.Subscriptions.Models;
-using Microsoft.Azure.PowerShell.Common.Config;
 using Microsoft.Identity.Client;
 using Microsoft.Identity.Client.Broker;
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Microsoft.Azure.Commands.Common.Authentication
 {
     public abstract class PowerShellTokenCacheProvider
     {
         public const string PowerShellTokenCacheProviderKey = "PowerShellTokenCacheProviderKey";
-        private static readonly string CommonTenant = "organizations";
+        //Refer to https://learn.microsoft.com/en-us/dotnet/api/microsoft.identity.client.abstractapplicationbuilder-1.withauthority?view=msal-dotnet-latest#microsoft-identity-client-abstractapplicationbuilder-1-withauthority(system-string-system-boolean
+        //However, neither "commons" nor "organizations" works for MSA account
+        private static readonly string OrganizationTenant = "organizations";
 
         protected byte[] _tokenCacheDataToFlush;
 
@@ -98,7 +98,7 @@ namespace Microsoft.Azure.Commands.Common.Authentication
                 Id = account.Username,
                 Type = AzureAccount.AccountType.User
             };
-            var commonToken = AzureSession.Instance.AuthenticationFactory.Authenticate(azureAccount, environment, CommonTenant, null, null, promptAction);
+            var commonToken = AzureSession.Instance.AuthenticationFactory.Authenticate(azureAccount, environment, OrganizationTenant, null, null, promptAction);
             IEnumerable<string> tenants = Enumerable.Empty<string>();
             using (SubscriptionClient subscriptionClient = GetSubscriptionClient(commonToken, environment))
             {
@@ -165,10 +165,9 @@ namespace Microsoft.Azure.Commands.Common.Authentication
         protected abstract void RegisterCache(IPublicClientApplication client);
 
         /// <summary>
-        /// Creates a public client app.
-        /// This method is not meant for authentication purpose. Use APIs from Azure.Identity instead.
+        /// Creates a public client app with tenantId.
         /// </summary>
-        public virtual IPublicClientApplication CreatePublicClient(string authority = null)
+        public virtual IPublicClientApplication CreatePublicClient(string authority, string tenantId = null)
         {
             var builder = PublicClientApplicationBuilder.Create(Constants.PowerShellClientId);
             if (AzConfigReader.IsWamEnabled(authority))
@@ -177,11 +176,19 @@ namespace Microsoft.Azure.Commands.Common.Authentication
             }
             if (!string.IsNullOrEmpty(authority))
             {
-                builder.WithAuthority(authority);
+                builder.WithAuthority(authority, tenantId ?? OrganizationTenant);
             }
             var client = builder.Build();
             RegisterCache(client);
             return client;
+        }
+        /// <summary>
+        /// Creates a public client app.
+        /// This method is not meant for authentication purpose. Use APIs from Azure.Identity instead.
+        /// </summary>
+        public virtual IPublicClientApplication CreatePublicClient(string authority = null)
+        {
+            return CreatePublicClient(authority);
         }
 
         public abstract TokenCachePersistenceOptions GetTokenCachePersistenceOptions();