+++ src/Accounts/Authentication/Factories/SshCredentialFactory.cs
@@ -69,7 +69,7 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
                 throw new NullReferenceException(Resources.AuthenticationClientFactoryNotRegistered);
             }
 
-            var publicClient = tokenCacheProvider.CreatePublicClient();
+            var publicClient = tokenCacheProvider.CreatePublicClient(context.Environment.ActiveDirectoryAuthority, context.Account.GetCommonTenant());
             string cloudName = context.Environment.Name.ToLower();
             string scope = CloudToScope.GetValueOrDefault(cloudName, null);
             if (scope == null)
@@ -79,7 +79,7 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
             List<string> scopes = new List<string>() { scope };
             var jwk = CreateJwk(rsaKeyInfo, out string keyId);
 
-            var account = publicClient.GetAccountAsync(context.Account.ExtendedProperties["HomeAccountId"])
+            var account = publicClient.GetAccountAsync(context.Account.GetProperty(AzureAccount.Property.HomeAccountId))
                             .ConfigureAwait(false).GetAwaiter().GetResult();
             var result = publicClient.AcquireTokenSilent(scopes, account)
                         .WithSSHCertificateAuthenticationScheme(jwk, keyId)